✨ [BE] 트렌딩 작품 리스트 기능 구현

See merge request s08-webmobile1-sub2/S08P12A108!25

+++ Backend/src/main/java/com/ssafy/banana/api/controller/ArtController.java
@@ -10,17 +10,21 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.nimbusds.oauth2.sdk.util.StringUtils;
 import com.ssafy.banana.api.service.ArtService;
 import com.ssafy.banana.db.entity.Art;
-import com.ssafy.banana.db.entity.MyArt;
 import com.ssafy.banana.dto.request.ArtRequest;
 import com.ssafy.banana.dto.request.MasterpieceRequest;
 import com.ssafy.banana.dto.request.MyArtRequest;
 import com.ssafy.banana.dto.response.ArtDetailResponse;
 import com.ssafy.banana.dto.response.ArtResponse;
+import com.ssafy.banana.exception.CustomException;
+import com.ssafy.banana.exception.CustomExceptionType;
+import com.ssafy.banana.security.jwt.TokenProvider;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -32,17 +36,21 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ArtController {
 
-	private static final Long TEST_USER_SEQ = 60L;
-
+	private static final String AUTHORIZATION = "Authorization";
+	private final TokenProvider tokenProvider;
 	private final ArtService artService;
 
 	@ApiOperation(value = "작품 업로드", notes = "나의 작품을 업로드합니다")
 	@PostMapping
-	public ResponseEntity uploadArt(@RequestBody ArtRequest artRequest) {
+	public ResponseEntity uploadArt(@RequestBody ArtRequest artRequest,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		Long userSeq = TEST_USER_SEQ;    // 수정 예정 (token으로 받아오기)
-		String artThumbnail = "artThumbnail 구해오기";    // 수정 예정
-		Art art = artService.uploadArt(artRequest, userSeq, artThumbnail);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
+		Art art = artService.uploadArt(artRequest, userSeq);
 
 		return ResponseEntity.status(HttpStatus.OK).body(art);
 	}
@@ -53,51 +61,74 @@ public class ArtController {
 
 		List<ArtResponse> artList = artService.getAllArtList();
 
-		if (artList.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
-		}
 		return ResponseEntity.status(HttpStatus.OK).body(artList);
+	}
 
+	@ApiOperation(value = "신규 작품 리스트", notes = "신규 작품 목록을 반환합니다")
+	@GetMapping("/new")
+	public ResponseEntity<List<ArtResponse>> getNewArtList() {
+
+		List<ArtResponse> artList = artService.getNewArtList();
+
+		return ResponseEntity.status(HttpStatus.OK).body(artList);
 	}
 
 	@ApiOperation(value = "나의 작품 리스트", notes = "작가 본인의 작품 목록을 반환합니다")
 	@GetMapping("/{user_seq}")
-	public ResponseEntity<List<ArtResponse>> getMyArtList(@PathVariable("user_seq") Long userSeq) {
+	public ResponseEntity<List<ArtResponse>> getMyArtList(@PathVariable("user_seq") Long userSeq,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		List<ArtResponse> artList = artService.getMyArtList(userSeq);
-
-		if (artList.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
 		}
-		return ResponseEntity.status(HttpStatus.OK).body(artList);
+		Long tokenUserSeq = tokenProvider.getSubject(token);
+		List<ArtResponse> artList = artService.getMyArtList(userSeq, tokenUserSeq);
 
+		return ResponseEntity.status(HttpStatus.OK).body(artList);
 	}
 
-	//    // 보류
-	//    @ApiOperation(value = "대표 작품 리스트", notes = "작가 본인의 대표작 목록을 반환합니다")
-	//    @GetMapping("/{user_seq}/masterpiece")
-	//    public ResponseEntity getMasterpieceList(@PathVariable("user_seq") Long userSeq){
-	//
-	//        return null;
-	//    }
+	@ApiOperation(value = "대표 작품 리스트", notes = "작가 본인의 대표작 목록을 반환합니다")
+	@GetMapping("/{user_seq}/masterpiece")
+	public ResponseEntity getMasterpieceList(@PathVariable("user_seq") Long userSeq,
+		@RequestHeader(AUTHORIZATION) String token) {
+
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long tokenUserSeq = tokenProvider.getSubject(token);
+		List<ArtResponse> artList = artService.getMasterpieceList(userSeq, tokenUserSeq);
+
+		return ResponseEntity.status(HttpStatus.OK).body(artList);
+	}
 
 	@ApiOperation(value = "좋아요한 작품 리스트", notes = "유저가 좋아요를 누른 작품 목록을 반환합니다")
 	@GetMapping("/{user_seq}/like")
-	public ResponseEntity<List<ArtResponse>> getLikedArtList(@PathVariable("user_seq") Long userSeq) {
+	public ResponseEntity<List<ArtResponse>> getLikedArtList(@PathVariable("user_seq") Long userSeq,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		List<ArtResponse> artList = artService.getLikedArtList(userSeq);
-		if (artList.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
 		}
+		Long tokenUserSeq = tokenProvider.getSubject(token);
+		List<ArtResponse> artList = artService.getLikedArtList(userSeq, tokenUserSeq);
 
 		return ResponseEntity.status(HttpStatus.OK).body(artList);
 	}
 
 	@ApiOperation(value = "대표 작품 설정", notes = "작가 본인의 대표작을 설정합니다")
 	@PutMapping("/masterpiece")
-	public ResponseEntity<?> setMasterpieceList(@RequestBody List<MasterpieceRequest> masterpieceRequestList) {
+	public ResponseEntity<?> setMasterpieceList(@RequestBody List<MasterpieceRequest> masterpieceRequestList,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		artService.setMasterpieceList(masterpieceRequestList);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
+		artService.setMasterpieceList(masterpieceRequestList, userSeq);
 
 		return ResponseEntity.status(HttpStatus.OK).body(null);
 	}
@@ -112,19 +143,14 @@ public class ArtController {
 		return ResponseEntity.status(HttpStatus.OK).body(artList);
 	}
 
-	// @ApiOperation(value = "트렌딩 작품 리스트", notes = "최근에 좋아요를 많이 받은 작품 목록을 반환합니다")
-	// @GetMapping("/trend")
-	// public ResponseEntity<List<ArtResponseDto>> getTrendArtList() {
-	//
-	// 	return null;
-	// }
-	//
-	// @ApiOperation(value = "새로 나온 작품 리스트", notes = "최근에 업로드 된 작품 목록을 반환합니다")
-	// @GetMapping("/new")
-	// public ResponseEntity<List<ArtResponseDto>> getNewArtList() {
-	//
-	// 	return null;
-	// }
+	@ApiOperation(value = "트렌딩 작품 리스트", notes = "최근 2주 동안에 좋아요를 많이 받은 작품 목록을 반환합니다")
+	@GetMapping("/trend")
+	public ResponseEntity<List<ArtResponse>> getTrendArtList() {
+
+		List<ArtResponse> artList = artService.getTrendArtList();
+
+		return ResponseEntity.status(HttpStatus.OK).body(artList);
+	}
 
 	@ApiOperation(value = "인기 작품 리스트", notes = "좋아요를 많이 받은 작품 목록을 반환합니다")
 	@GetMapping("/popular")
@@ -132,16 +158,22 @@ public class ArtController {
 
 		List<ArtResponse> artList = artService.getPopularArtList();
 
-		if (artList.isEmpty()) {
-			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
-		}
 		return ResponseEntity.status(HttpStatus.OK).body(artList);
 	}
 
 	@ApiOperation(value = "작품 상세 정보", notes = "작품의 상세 정보를 반환합니다")
 	@GetMapping("/detail/{art_seq}")
-	public ResponseEntity getArt(@PathVariable("art_seq") Long artSeq) {
+	// @PreAuthorize("hasRole('USER')")
+	public ResponseEntity getArt(@PathVariable("art_seq") Long artSeq,
+		@RequestHeader(value = AUTHORIZATION, required = false) String token) {
 
+		if (StringUtils.isBlank(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
 		ArtDetailResponse artDetailResponse = artService.getArt(artSeq);
 
 		return ResponseEntity.status(HttpStatus.OK).body(artDetailResponse);
@@ -149,12 +181,32 @@ public class ArtController {
 
 	@ApiOperation(value = "작품 좋아요 추가하기", notes = "작품에 좋아요를 설정합니다")
 	@PostMapping("/like")
-	public ResponseEntity addArtLike(@RequestBody MyArtRequest myArtRequest) {
+	public ResponseEntity addArtLike(@RequestBody MyArtRequest myArtRequest,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		Long userSeq = TEST_USER_SEQ;    // 수정 예정 (token으로 받아오기)
-		MyArt myArt = artService.addArtLike(myArtRequest, userSeq);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
+		Art art = artService.addArtLike(myArtRequest, userSeq);
 
-		return ResponseEntity.status(HttpStatus.OK).body(myArt);
+		return ResponseEntity.status(HttpStatus.OK).body(art);
+	}
+
+	@ApiOperation(value = "작품 좋아요 삭제하기", notes = "작품에 좋아요를 취소합니다")
+	@DeleteMapping("/like")
+	public ResponseEntity deleteArtLike(@RequestBody MyArtRequest myArtRequest,
+		@RequestHeader(AUTHORIZATION) String token) {
+
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
+		Art art = artService.deleteArtLike(myArtRequest, userSeq);
+
+		return ResponseEntity.status(HttpStatus.OK).body(art);
 	}
 
 	// @ApiOperation(value = "작품 다운로드", notes = "작품을 다운로드합니다")
@@ -166,26 +218,39 @@ public class ArtController {
 
 	@ApiOperation(value = "작품 수정", notes = "등록된 작품을 수정합니다")
 	@PutMapping
-	public ResponseEntity modifyArt(@RequestBody ArtRequest artRequest) {
+	public ResponseEntity updateArt(@RequestBody ArtRequest artRequest,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		Long userSeq = TEST_USER_SEQ;    // 수정 예정 (token으로 받아오기)
-		String artThumbnail = "artThumbnail 구해오기";    // 수정 예정
-		Art art = artService.updateArt(artRequest, userSeq, artThumbnail);
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
+		Art art = artService.updateArt(artRequest, userSeq);
 
 		return ResponseEntity.status(HttpStatus.OK).body(art);
 	}
 
 	@ApiOperation(value = "작품 삭제", notes = "등록된 작품을 삭제합니다")
-	@DeleteMapping("art_seq")
-	public ResponseEntity deleteArt(@PathVariable("art_seq") Long artSeq) {
+	@DeleteMapping("/{art_seq}")
+	public ResponseEntity deleteArt(@PathVariable("art_seq") Long artSeq,
+		@RequestHeader(AUTHORIZATION) String token) {
 
-		Long userSeq = TEST_USER_SEQ;    // 수정 예정 (token으로 받아오기)
+		token = getToken(token);
+		if (!tokenProvider.validateToken(token)) {
+			throw new CustomException(CustomExceptionType.AUTHORITY_ERROR);
+		}
+		Long userSeq = tokenProvider.getSubject(token);
 		Long result = artService.deleteArt(artSeq, userSeq);
 
-		if (result == -1L) {
-			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+		return ResponseEntity.status(HttpStatus.OK).body(result);
+	}
+
+	private static String getToken(String token) {
+		if (token.substring(0, 7).equals("Bearer ")) {
+			token = token.substring("Bearer ".length());
 		}
-		return ResponseEntity.status(HttpStatus.OK).body(null);
+		return token;
 	}
 
 }
